# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AI Agent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /v1/users/me/password:
    patch:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- Projects ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: ["object", "null"]
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth schemas
...
    User:
      type: object
      required: [id, username, email, role, is_active, created_at, updated_at]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar_url:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, moderator, admin, super_admin]
        is_active:
          type: boolean
        last_login:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        avatar_url:
          type: ["string", "null"]
    ChangePasswordInput:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8

    # Project schemas
...